requirements.txt
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
cffi==1.15.0
cryptography==36.0.0
decorator==5.1.0
Django==4.0
django-extensions==3.1.5
django-filter==21.1
django-test-plus==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
iniconfig==1.1.1
ipython==7.30.1
jedi==0.18.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
